+++ llama_index/evaluation/__init__.py
@@ -1,13 +1,6 @@
 """Evaluation modules."""
 
-from llama_index.evaluation.answer_consistency import AnswerConsistencyEvaluator
-from llama_index.evaluation.answer_consistency_binary import (
-    AnswerConsistencyBinaryEvaluator,
-)
 from llama_index.evaluation.answer_relevancy import AnswerRelevancyEvaluator
-from llama_index.evaluation.answer_similarity import AnswerSimilarityEvaluator
-from llama_index.evaluation.augmentation_accuracy import AugmentationAccuracyEvaluator
-from llama_index.evaluation.augmentation_precision import AugmentationPrecisionEvaluator
 from llama_index.evaluation.base import (
     BaseEvaluator,
     EvaluationResult,
@@ -38,9 +31,28 @@ from llama_index.evaluation.retrieval.metrics import (
     RetrievalMetricResult,
     resolve_metrics,
 )
-from llama_index.evaluation.retrieval_precision import RetrievalPrecisionEvaluator
 from llama_index.evaluation.semantic_similarity import SemanticSimilarityEvaluator
-from llama_index.evaluation.tonic_validate_evaluator import TonicValidateEvaluator
+from llama_index.evaluation.tonic_validate.answer_consistency import (
+    AnswerConsistencyEvaluator,
+)
+from llama_index.evaluation.tonic_validate.answer_consistency_binary import (
+    AnswerConsistencyBinaryEvaluator,
+)
+from llama_index.evaluation.tonic_validate.answer_similarity import (
+    AnswerSimilarityEvaluator,
+)
+from llama_index.evaluation.tonic_validate.augmentation_accuracy import (
+    AugmentationAccuracyEvaluator,
+)
+from llama_index.evaluation.tonic_validate.augmentation_precision import (
+    AugmentationPrecisionEvaluator,
+)
+from llama_index.evaluation.tonic_validate.retrieval_precision import (
+    RetrievalPrecisionEvaluator,
+)
+from llama_index.evaluation.tonic_validate.tonic_validate_evaluator import (
+    TonicValidateEvaluator,
+)
 
 # import dataset generation too
 from llama_index.finetuning.embeddings.common import (
